+++ core/src/main/scala/org/broadinstitute/dsp/workbench/welder/Crc32c.scala
@@ -5,6 +5,8 @@ import java.util.zip.Checksum
 import cats.effect.{ContextShift, IO}
 import com.google.common.io.BaseEncoding
 import org.broadinstitute.dsde.workbench.google2.Crc32
+
+import scala.concurrent.ExecutionContext
 import scala.concurrent.ExecutionContext.global
 // Adapted from https://github.com/googlearchive/crc32c-java/blob/master/src/com/google/cloud/Crc32c.java, which is the only java implementation for google's crc32c algorithm.
 // Note: The CRC popularly known as CRC32 differs from the CRC32c algorithm used by Cloud Storage.
@@ -38,7 +40,7 @@ object Crc32c {
   private val LONG_MASK = 0XFFFFFFFFL
   private val BYTE_MASK = 0xff
 
-  def calculateCrc32ForFile(path: java.nio.file.Path)(implicit cs: ContextShift[IO]): IO[Crc32] = {
+  def calculateCrc32ForFile(path: java.nio.file.Path, blockingEc: ExecutionContext)(implicit cs: ContextShift[IO]): IO[Crc32] = {
     fs2.io.file.readAll[cats.effect.IO](path, global, 4096).compile.to[Array].map(bytes => calculateCrc32c(bytes))
   }
